* Formatting changes

+++ src/Grid.h
@@ -19,8 +19,8 @@ public:
 
 	ImColor GetCellMainCellColour();
 
-	float GetGridSteps() const;
-	void SetGridSteps(unsigned int);
+	int GetGridSteps() const;
+	void SetGridSteps(int);
 
 	void SetMainCellColour(ImColor);
 
@@ -44,7 +44,7 @@ private:
 	ImVec2 m_max_canvas_position;
 	
 	ImVec2 m_grid_scrolling;
-	unsigned int m_grid_steps;
+	int m_grid_steps;
 
 	ImColor m_cell_colour_main;
